--- Web/main.py
@@ -1,14 +1,0 @@
-from flask import Flask
-from flask import render_template
-app = Flask(__name__) 
-
-@app.route('/')
-def  base():    
-     return render_template("base.html")
-
-@app.route('/cart')
-def cart():
-      return render_template("cart.html")
-
-if __name__=="__main__":
-       app.run(debug=True,host='0.0.0.0')
